+++ pkg/hostmeta/read_access_other.go
@@ -1,0 +1,9 @@
+//go:build !darwin
+
+package hostmeta
+
+import "os"
+
+func recordReadAccess(_ *os.File) error {
+	return ErrReadAccessUnsupported
+}